Create a session to test zero down time deployments
In order to support a zero down time deployment we need to be sure that
the old deployment is allowed to cease operation before any cut over is
made.

Actually create an HTTP session to facilitate this.  A next step might be
to make this configurable.

For now just set a 5 minute timeout.